pyproject.toml
[tool.poetry]
name = "get-papers-list"
version = "0.1.0"
description = "Fetch PubMed papers with non-academic authors"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{ include = "get_papers_list" }]

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.32.0"
llama-cpp-python = "^0.3.12"
tqdm = "^4.66.0"
ftfy = "^6.3.1"
googletrans = "4.0.0rc1"

[tool.poetry.scripts]
get-papers-list = "get_papers_list.cli:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
